Add parallel writing for reference
This is the solution for "Checkpoints" up to 2.2 later stages of the
task remove the parallel writing process to guarantee the atomicity of
the write operation. Another solution would be to create a temporary
file. But this needs again more write operations. If the centralized or
parallel in this case is better would be left to the benchmarks. For
simplicity and consistency a buffer is integrated in main memory in our
code.