Add solution "Checkpoints" (#9)
* Implement first two subtasks

* Continue calculating from last iteration

* Add atomic writing

* Add missing barrier

Forgot this one in the previous commit, otherwise the write results are
scrambled.

* Add parallel writing for reference

This is the solution for "Checkpoints" up to 2.2 later stages of the
task remove the parallel writing process to guarantee the atomicity of
the write operation. Another solution would be to create a temporary
file. But this needs again more write operations. If the centralized or
parallel in this case is better would be left to the benchmarks. For
simplicity and consistency a buffer is integrated in main memory in our
code.

* Add additional answers to checkpoints

* Add disclaimer for copied source file

* Use default name in case none is given

* Add statistics

* Fix misc

* Update parallel impl

* Fix some linter errors

* Add final answer